Added GUI, enhanced CLI, and made python3
When launching the GenSMBIO.command or .bat, It now checks for python 3, and prompts to install if needed. It also creates a .venv to not cause any problems with the rest of the system.

Modenized the code to Python 3. No more python 2. Python 2 was sunseted over 5 years ago.

Now wuns in gui by default, but can still use text interface with a --tui, or is tkinter is not installed. like in a linux terminal.